refactor: constants 파일을 역할에 따라 분리
