+++ src/constants/carMovementConstants.js
@@ -1,0 +1,2 @@
+export const MIN_MOVEMENT = 4;
+export const MOVEMENT_FORWARD = '-';
