+++ src/constants/carNameConstants.js
@@ -1,0 +1,2 @@
+export const MIN_NAME_LENGTH = 1;
+export const MAX_NAME_LENGTH = 5;
