--- src/constants/constants.js
@@ -1,11 +1,0 @@
-// car name length
-export const MIN_NAME_LENGTH = 1;
-export const MAX_NAME_LENGTH = 5;
-
-// random number generation
-export const MIN_RANGE = 0;
-export const MAX_RANGE = 9;
-
-// car movement
-export const MIN_MOVEMENT = 4;
-export const MOVEMENT_FORWARD = '-';
