+++ src/constants/randomNumberConstants.js
@@ -1,0 +1,2 @@
+export const MIN_RANGE = 0;
+export const MAX_RANGE = 9;
